refactor: text 토큰 적용
